# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sito DNS Administrative API",
    "description": "Administrative REST API and OpenAPI 3.0 specification for the sito DNS server",
    "license": {
      "name": "GPL-3.0-only"
    },
    "version": "1.6.0"
  },
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Get current full configuration with secrets masked.",
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "Full configuration with masked secrets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Update full configuration atomically after pre-commit validation.",
        "operationId": "update_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/config/backup": {
      "get": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Download a complete configuration backup archive (.tar.gz).",
        "operationId": "download_backup",
        "responses": {
          "200": {
            "description": "Gzipped tar archive containing config and metadata",
            "content": {
              "application/gzip": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/config/reload": {
      "post": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Reload configuration from disk without server restart.",
        "operationId": "reload_config",
        "responses": {
          "200": {
            "description": "Configuration reloaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration on disk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/config/restore": {
      "post": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Upload and validate a backup archive (.tar.gz), generating a confirmation token.",
        "operationId": "prepare_restore",
        "requestBody": {
          "description": "Gzipped tar archive",
          "content": {
            "application/gzip": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Backup validated and confirmation token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestorePreparedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Corrupted or invalid backup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/config/restore/confirm": {
      "post": {
        "tags": [
          "crate::handlers::config"
        ],
        "summary": "Confirm and apply configuration restoration using confirmation token.",
        "operationId": "confirm_restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Restoration completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired confirmation token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigResponse": {
        "type": "object",
        "description": "Full configuration response with secrets masked.",
        "required": [
          "config_toml"
        ],
        "properties": {
          "config_toml": {
            "type": "string"
          }
        }
      },
...
      "GenericMessageResponse": {
        "type": "object",
        "description": "Generic success message response.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response object.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Detailed explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional URI reference identifying the specific occurrence of the problem."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary of the problem."
          },
          "type": {
            "type": "string",
            "description": "URI reference identifying the problem type."
          }
        }
      },
...
      "RestoreConfirmRequest": {
        "type": "object",
        "description": "Request to confirm restoration.",
        "required": [
          "confirmation_token"
        ],
        "properties": {
          "confirmation_token": {
            "type": "string"
          }
        }
      },
      "RestorePreparedResponse": {
        "type": "object",
        "description": "Restore preparation response with confirmation token.",
        "required": [
          "confirmation_token",
          "message",
          "config_preview"
        ],
        "properties": {
          "config_preview": {
            "type": "string"
          },
          "confirmation_token": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateConfigRequest": {
        "type": "object",
        "description": "Request to replace full configuration.",
        "required": [
          "config_toml"
        ],
        "properties": {
          "config_toml": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}